% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mosoblwithadd.R
\name{combine}
\alias{combine}
\title{It's plot moskovskaya oblast'}
\usage{
combine(layer, add, n, clr)
}
\arguments{
\item{layer}{- the initial SpatialPolygonsDataFrame file.

add - vector of additional SpatialPolygonsDataFrame file. The size of each image should
be comparable with size of the initial image.

n - number of additional layers

clr - vector with colours of all additional images.}
}
\value{
Image of mosobl with addings.
}
\description{
Combination of two or more layers of information on one map.
}
\details{
Map of Moskovskaya oblast' with railway lines.
}
\examples{
mosobl <- readOGR("mosobl", "boundary-polygon")
railway_mo <- readOGR("mosobl", "railway-line")
waterline_mo <- readOGR("mosobl", "water-line")
additional_files <- c(railway_mo, waterline_mo)
colour = c("red", "blue")
combine(layer = mosobl, add = additional_files, n = 2, clr = colour)
}

