% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmap.R
\name{findRegbyID}
\alias{findRegbyID}
\title{Find regyons by ID}
\usage{
findRegbyID(vector, layer)
}
\arguments{
\item{vec}{- vector with IDs of regions,
layer - SpatialPolygonsDataFrame file,
colour - colour of region}
}
\value{
Image of map with colourful selected regions
}
\description{
This function looking for regyons, defined by the ID, at the map.
}
\details{
Built several regyons by there IDs.
}

