% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmapwithcolour.R
\name{ggplotcolby}
\alias{ggplotcolby}
\title{It's colour map by some parameter}
\usage{
ggplotcolby(layer, data, par = "POP_2002")
}
\arguments{
\item{par}{- chosen parameter,
layer - SpatialPolygonsDataFrame file}
}
\value{
Image of map with
}
\description{
Colouring by parameter
}
\details{
Map with graduate colour
}
\examples{
map <- readOGR(dsn = "data", layer = "map_1")
pop <- read.csv("csv/population.csv", sep = ';', encoding = "UTF-16")
ggplotcolby(layer = map, data = pop, par = "POP_2002")
}

