% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmapwithcolour.R
\name{plotcolby}
\alias{plotcolby}
\title{It's colour map by some parameter}
\usage{
plotcolby(layer, data, col = "gold", par = "POP_2002")
}
\arguments{
\item{layer}{- SpatialPolygonsDataFrame file

data  - csv file with data.

col  - colour of map.

par - chosen parameter, as in data file.}
}
\value{
Image of map, shaded by choosen parameter.
}
\description{
Colouring by parameter
}
\details{
Map with graduate colouring.
}
\examples{
map <- readOGR(dsn = "data", layer = "map_1")
pop <- read.csv("csv/population.csv", sep = ';', encoding = "UTF-16")
plotcolby(layer = map, data = pop, col = "gold", par = "POP_2002")
}

