% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmap.R
\name{plotRegbyID}
\alias{plotRegbyID}
\title{It's plot single region by it's ID}
\usage{
plotRegbyID(n, layer)
}
\arguments{
\item{n}{- region's ID number,
layer - SpatialPolygonsDataFrame file}
}
\value{
Image of region
}
\description{
Plotting region by ID
}
\details{
Build part of map by ID
}
\examples{
map <- readOGR(dsn = "data", layer = "CorCoor")
plotRegbyID(n=1, layer = map)
}

