% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmapwithmatrix.R
\name{reg.matrix.small}
\alias{reg.matrix.small}
\title{Several chosen regions with smaller administrarive division}
\usage{
reg.matrix.small(layer_2, regions, outname = "matrix_with_regions.jpeg",
  q = 60000, x = 3, y = 3, t = 2)
}
\arguments{
\item{layer_2}{- SpatialPolygonsDataFrame file with smaller administrative division.
regions - a character vector naming one of more regions, as in layer_2.
outname - name of JPEG file.
q - quality of image.
x - number of images in the vertical.
y - number of images in the horizontal.
t - size of text.}
}
\value{
JPEG file with chosen regions.
}
\description{
Computes the chosen regions in JPEG file
}
\details{
Each region will be imaged in single cell of matrix of a given size.
}
\examples{
map <- readOGR(dsn = "data", layer_2 = "map_2")
regions.matrix(layer_2 = map_2, regions = c("Tuva", "Tomsk"), outname = "First.jpeg")
regions.matrix(layer_2 = map_2, regions = c("Tuva", "Tomsk"), outname = "Second.jpeg", q = 50)
regions.matrix(layer_2 = map_2, regions = c("Tuva", "Tomsk"), outname = "Third.jpeg", q = 200, x = 10)
}

