% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmap.R
\name{colby}
\alias{colby}
\title{It's colour map by some parameter}
\usage{
colby(par, layer_f)
}
\arguments{
\item{par}{- chosen parameter,
layer - SpatialPolygonsDataFrame file}
}
\value{
Image of map with
}
\description{
Colouring by parameter
}
\details{
Map with graduate colour
}

