% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmapwithcolour.R
\name{ggplotcolby}
\alias{ggplotcolby}
\title{It's colour map by some parameter}
\usage{
ggplotcolby(layer, data, col = "gold", par = "POP_2002")
}
\arguments{
\item{par}{- chosen parameter,
layer - SpatialPolygonsDataFrame file}
}
\value{
Image of map with
}
\description{
Colouring by parameter
}
\details{
Map with graduate colour
}
\examples{
map <- readOGR(dsn = "data", layer = "map_1")
ggplotcolby(par = ID_1, layer = map)
}

